% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesforcerExtendr.R
\docType{package}
\name{salesforcerExtendr}
\alias{salesforcerExtendr}
\alias{salesforcerExtendr-package}
\title{Add ons for the excellent salesforcer package}
\description{
A collection of additional helpers functions to extend the salesforcer package
}
\author{
Mark Goble \email{mark.goble@goble.co.uk}
}
