% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_bind_object_fields.R
\name{sf_bind_object_fields}
\alias{sf_bind_object_fields}
\title{Binds a salesforce object to it's fields}
\usage{
sf_bind_object_fields(objects_data_frame, fields_data_frame)
}
\arguments{
\item{objects_data_frame}{a \code{data frane} of object information}

\item{fields_data_frame}{a \code{data frame} of the field information}
}
\description{
This helper function will bind the fields to their object descriptions to get the final data frame.
 For internal use within the package
}
\details{
Notes and Limitations:
none
}
\examples{
\dontrun{
to add
}
}
