% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_get_object_details.R
\name{sf_get_object_details}
\alias{sf_get_object_details}
\title{Returns details of objects / fields in an org}
\usage{
sf_get_object_details(use_connection_file = FALSE,
  connection_file = NULL, directory = NULL)
}
\arguments{
\item{use_connection_file}{should a connection file be used? (set to false if a session has already been established)}

\item{connection_file}{only used if use-connection_file is TRUE. Passed to Secure_connection}

\item{directory}{only used if use_connection_file is TRUE. Passed to Secure_connection}
}
\value{
A \code{list} of two data frames (see intoduction)
}
\description{
This function will return a list of two data frames - one with fields from objects
which can be seen under the users permissions (populated objects) and the other
for objects which don't have any visible fields (unpopulated objects)
}
\details{
If use connection file is true then uses the secure_connection method otherwise
will open salesforce in a browser to authenticate.
}
\examples{
\dontrun{
sf_get_object_details()
}
}
