% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_object_fields.R
\name{sf_object_fields}
\alias{sf_object_fields}
\title{Gets all the fields for a salesforce object}
\usage{
sf_object_fields(sf_object_name)
}
\arguments{
\item{sf_object_name}{the name of the salesforce object to query}
}
\value{
A \code{data frame} of the fields in the object or 'empty' if no readable fields
}
\description{
This function will pull back the information for a named Object.
}
\details{
Notes and Limitations:

Note: this function can actually pull back info for objects such as 'Activity'
which aren't Sobjects.

This version doesn't retrieve some field details for standard fields - for
example the data type. This is down to the way that the Force.com query works.
The SF query doesn't return the 'created' and 'modified' fields.

Also note that valueset (which contains picklist values) and
summaryfilteritems are also removed as these are nested lists.

for v2 changed from 'CustomObject' to 'CustomDatatyoeComponent' as we don't
need all the data which is returned under 'customObject'
}
\examples{
\dontrun{
sf_object_fields("contact")
}
}
