% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_objects_info.R
\name{sf_objects_info}
\alias{sf_objects_info}
\title{Gets all salesfroce objects from an org}
\usage{
sf_objects_info()
}
\value{
A \code{data frame} of objects and their fields
}
\description{
This function will pull back the information on all sObjects in the salesforce org.
}
\details{
Notes and Limitations

Objects such as Activity are not returned as these aren't Sobjects (they
are more akin to views). Also note that the 'URLs' are stripped from out as
these are nested and also not required. The function will only return objects
which it has permission to see.

Requires a connection to be established to salesforce.
}
\examples{
\dontrun{
sf_objects_info()
}
}
