% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_objects_iteration.R
\name{sf_objects_iteration}
\alias{sf_objects_iteration}
\title{oterates over a collection of salesforce objects to get the fields for each}
\usage{
sf_objects_iteration(objects_data_frame, name_column = "name")
}
\arguments{
\item{objects_data_frame}{the data frame containing the object list}

\item{name_column}{the column containing the names of the objects}
}
\value{
A \code{list} of objects and their fields
}
\description{
This function will iterate over each object to get the object fields.
}
\details{
Notes and Limitations

See also sf_object_fields. This creates a list containing an entry for each
queried object. If sf_object_fields doesn't return any info (for
example if all the fields are hidden / not return) then the list entry will be
"empty".
}
\examples{
\dontrun{
sf_objects_iteration(sobjects)
}
}
