% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_unpopulated_objects.R
\name{sf_unpopulated_objects}
\alias{sf_unpopulated_objects}
\title{creates a data frame of unpopulated salesforce objects}
\usage{
sf_unpopulated_objects(objects_fields_list, objects_data_frame)
}
\arguments{
\item{objects_fields_list}{a \code{list} of the fields of the objects}

\item{objects_data_frame}{a \code{data frame} of the details of the objects}
}
\value{
a \code{data frame} of all the unpopulated objects and their fields
}
\description{
This function will reformat the data from an object list into a data frame.
}
\details{
Notes and Limitations:

Takes the output and filter to just get those list items which DON'T contain
object details
}
\examples{
\dontrun{
to add
}
}
